+++ vortex-ffi/src/array.rs
@@ -422,4 +422,41 @@ mod tests {
             vx_array_free(ffi_array);
         }
     }
+
+    #[test]
+    fn test_array_dtype_lifetime_pattern() {
+        use vortex::arrays::StructArray;
+        use vortex::buffer::Buffer;
+        use vortex::{IntoArray, ArrayRef};
+        use crate::dtype::{vx_dtype_get_variant, vx_dtype_variant};
+        
+        // Helper function to create test array
+        fn create_test_struct_array() -> ArrayRef {
+            let nums: Buffer<i32> = (0..1000).collect();
+            let floats: Buffer<f32> = (0..1000).map(|x| x as f32).collect();
+            
+            StructArray::try_from_iter([
+                ("nums", nums.into_array()), 
+                ("floats", floats.into_array())
+            ])
+            .unwrap()
+            .into_array()
+        }
+        
+        let array = create_test_struct_array();
+        let vx_arr = vx_array::new(array);
+        
+        // Get dtype reference - this is valid as long as array lives
+        let dtype_ptr = unsafe { vx_array_dtype(vx_arr) };
+        let variant = unsafe { vx_dtype_get_variant(dtype_ptr) };
+        assert_eq!(variant, vx_dtype_variant::DTYPE_STRUCT);
+        
+        // Proper usage: use dtype while array is still alive
+        // This demonstrates the correct lifetime pattern
+        
+        unsafe { vx_array_free(vx_arr) };
+        
+        // Note: dtype_ptr is now invalid - this test documents the lifetime pattern
+        // In real usage, don't access dtype_ptr after freeing the array
+    }
 }
