+++ vortex-ffi/src/lib.rs
@@ -19,10 +19,6 @@ mod sink;
 mod string;
 mod struct_fields;
 
-#[cfg(test)]
-mod dtype_test;
-#[cfg(test)]
-mod memory_audit_test;
 
 use std::ffi::{CStr, c_char, c_int};
 use std::sync::Arc;
